+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log
 
-ver 1.0.1
+ver 0.3
 
 
-### 1.0.0
+### 0.3
 
-+ Initial release
++ unpack mode from export JSON file to files
+
+### 0.2
+
++ pack menus, blocks, langs, tables, params section from files to export JSON file
+
+### 0.1
+
++ pack contracts and pages section from files to export JSON file